Moving menu to resources, excising the maps for now
