+++ hendrix_today_app/pubspec.yaml
@@ -104,9 +104,6 @@ flutter:
   # example:
 
   fonts:
-      - family: Lora 
-        fonts:
-          - asset: fonts/Lora-Regular.ttf
       - family: Museo 
         fonts:
           - asset: fonts/Museo_Slab_300.otf
